+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
 import 'package:firebase_core/firebase_core.dart';
 import 'package:untitled/splashpage.dart';
 import 'firebase_options.dart';
