+++ lib/memo_insert.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:untitled/practice.dart';
 import 'package:untitled/setting.dart';
 import 'package:untitled/test.dart';
+import 'package:untitled/util.dart';
 
 import 'memo.dart';
