+++ lib/test.dart
@@ -280,44 +280,3 @@ class MakeTest extends StatelessWidget {
     );
   }
 }
-
-class Problem_t {
-  int page;
-  int number;
-  String title;
-  String testCase;
-  String rtestCase2;
-  String contents;
-  String myAnswer;
-  String hint;
-  String answer;
-
-  //Problem(몇장, 몇번문제, 문제제목, 소제목, 내용1, 입력받는 정답 ,내용2, 힌트, 정답);
-  static List<Problem_t> problems = [
-    Problem_t(1, 1,
-        '1-1\n<"Hi C programing">\n을 출력하세요.',
-        '10\n20\n30\n40\n50\n60',
-        'Hi C programing',
-        '#include <stdio.h>\n'
-        'int main(){\n'
-        '//안에 들어갈 내용을 적으세요.\n}',
-        '',
-        'hint',
-        'answer'),
-  ];
-
-  static Problem_t get(int a, int b) {
-    for (Problem_t p in problems) {
-      if (a == p.page && b == p.number) {
-        return p;
-      }
-    }
-    return dummy;
-  }
-
-  static Problem_t dummy = Problem_t(0, 0, 'dummy title', 'dummy testcase','dummy rtestcase',
-      'contents', 'dummy', 'hint', 'answer');
-
-  Problem_t(this.page, this.number, this.title, this.testCase, this.rtestCase2, this.contents,
-      this.myAnswer, this.hint, this.answer);
-}
